PackageBuilding.md: make building less idealistic
While there are many things one can do, in the team we mostly use
buildpackage and sbuild - let us document that as the chosen path
and simplify a lot due to that.

Furthermore we lacked the sbuild example to olink to the setup of it
as well as showing the actual long commandline that you'd use in most
cases.

And finally in doing so explain the arguments used, so this page
can be better for learning as you encounter it working in the server
team.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>