MergeProposalReview.md: add the option to skip a check
We define more and more checks, those make sense but there are cases
where particular checks would be a waste of time. Yet "not applicable"
is not the right status for it - therefore let us define a "skipped"
state which we can use when we consider a particular check not
commensurate.

This way we can more freely add tests that only apply to a subset,
helping us to not forget them while not making it a burden in general.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>