MergeProposalReview.md: do not explain two times
Due to the way how this evolved we had an intro and a template.
Then later added explanations and extended the template.
Due to that the intro was just a weaker form of the explanations
causing redundancy and mismatches. Replace the intro with
only a high level outline of what the explanations will then
cover in detail and add why we do this to be a true intro.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>