VersionStrings: Copyedit introductory text
This mainly focuses on revising the text for clarity, but I rewrote the
explanation of the debian revision to be more consistent with the spec.
"debian-revision" technically should refer to everything after the dash,
not just the debian specific number.  I introduce the terminology
'component' to help diffentiate between the debian number and the ubuntu
number in the revision.